% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{add_secret_version}
\alias{add_secret_version}
\title{Add a Secret Version}
\usage{
add_secret_version(project_id, secret_id, payload, etag = NULL)
}
\arguments{
\item{project_id}{The Google Cloud Project ID.}

\item{secret_id}{The ID of the secret to which the version will be added.}

\item{payload}{The secret data (as a raw vector or character string).}

\item{etag}{Optional. For optimistic concurrency control.}
}
\value{
A list containing information about the newly created secret version.
}
\description{
Adds a new version to an existing secret.
}
\examples{
\dontrun{
# Add a new version with string data
add_secret_version(
  project_id = "your-gcp-project-id",
  secret_id = "my-secret",
  payload = "my-new-secret-value"
)

# Add a new version with raw data
add_secret_version(
  project_id = "your-gcp-project-id",
  secret_id = "my-secret",
  payload = charToRaw("binary_data")
)
}
}
