% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{as_secret.character}
\alias{as_secret.character}
\title{As Secret Character}
\usage{
\method{as_secret}{character}(x, ...)
}
\arguments{
\item{x}{A string}

\item{...}{Unused argument needed for method}
}
\description{
As Secret Character
}
