% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secrets.R
\name{print.secret}
\alias{print.secret}
\title{Secret Print Method}
\usage{
\method{print}{secret}(x, ...)
}
\arguments{
\item{x}{A secret}

\item{...}{Unused argument required for method}
}
\description{
Secret Print Method
}
