% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{print.sm_secret}
\alias{print.sm_secret}
\title{Print sm_secret}
\usage{
\method{print}{sm_secret}(x, ...)
}
\arguments{
\item{x}{A sm_secret object}

\item{...}{Additional arguments passed to method}
}
\description{
Print sm_secret
}
