% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{print.sm_tbl}
\alias{print.sm_tbl}
\title{Print sm_tbl}
\usage{
\method{print}{sm_tbl}(x, ...)
}
\arguments{
\item{x}{A sm_tbl object}

\item{...}{Additional arguments passed to method}
}
\description{
Print sm_tbl
}
