% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secretmanager_endpoint}
\alias{secretmanager_endpoint}
\title{Secret Manager Endpoint}
\usage{
secretmanager_endpoint(i)
}
\arguments{
\item{i}{The index of the endpoint}
}
\description{
Secret Manager Endpoint
}
