% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{secretmanager_endpoints}
\alias{secretmanager_endpoints}
\alias{secretmanager_endpoint}
\title{List Secret Manager endpoints}
\usage{
secretmanager_endpoints(i = NULL)

secretmanager_endpoint(i)
}
\arguments{
\item{i}{The name(s) or integer index(ices) of the endpoints to return. \code{i}
is optional for \code{secretmanager_endpoints()} and, if not given, the entire list is
returned.}
}
\value{
One or more of the Secret Manager API v1 endpoints that are used internally by
secretmanager.
}
\description{
The secretmanager package stores a named list of Secret Manager API v1 endpoints (or
"methods", using Google's vocabulary) internally and these functions expose
this data.
\itemize{
\item \code{secretmanager_endpoint()} returns one endpoint, i.e. it uses \code{[[}.
\item \code{secretmanager_endpoints()} returns a list of endpoints, i.e. it uses \code{[}.
}

The names of this list (or the \code{id} sub-elements) are the nicknames that can
be used to specify an endpoint in \code{\link[=request_generate]{request_generate()}}. For each endpoint, we
store its nickname or \code{id}, the associated HTTP verb, the \code{path}, and details
about the parameters. This list is derived programmatically from the Secret Manager API v1 Discovery Document
(\verb{https://www.googleapis.com/discovery/v1/apis/secretmanager/v1/rest}) using the
approach described in the \href{https://gargle.r-lib.org/articles/request-helper-functions.html#discovery-documents}{Discovery Documents section}
of the gargle vignette \href{https://gargle.r-lib.org/articles/request-helper-functions.html}{Request helper functions}.
}
\examples{
str(head(secretmanager_endpoints(), 3), max.level = 2)
secretmanager_endpoint("secretmanager.projects.secrets.versions.destroy")
secretmanager_endpoint(4)
}
