% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{secretmanager_has_token}
\alias{secretmanager_has_token}
\title{Check if a token is available}
\usage{
secretmanager_has_token()
}
\value{
Logical. \code{TRUE} if a token is available, \code{FALSE} otherwise.
}
\description{
Checks if a token is present in the internal authentication state.
This is an internal helper function.
}
\keyword{internal}
