% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sm_endpoints}
\alias{sm_endpoints}
\title{List Secret Manager endpoints}
\usage{
sm_endpoints(i = NULL)
}
\arguments{
\item{i}{The name(s) or integer index(ices) of the endpoints to return. \code{i}
is optional for \code{sm_endpoints()} and, if not given, the entire list is
returned.}
}
\value{
One or more of the Secret Manager API v1 endpoints that are used internally by
googlesecretmanager.
}
\description{
The googlesecretmanager package stores a named list of Secret Manager API v1 endpoints (or
"methods", using Google's vocabulary) internally and these functions expose
this data.
\itemize{
\item \code{sm_endpoint()} returns one endpoint, i.e. it uses \code{[[}.
\item \code{sm_endpoints()} returns a list of endpoints, i.e. it uses \code{[}.
}

The names of this list (or the \code{id} sub-elements) are the nicknames that can
be used to specify an endpoint in request_generate(). For each endpoint, we
store its nickname or \code{id}, the associated HTTP verb, the \code{path}, and details
about the parameters. This list is derived programmatically from the Secret Manager API v1 Discovery Document
(\verb{https://www.googleapis.com/discovery/v1/apis/secretmanager/v1/rest}) using the
approach described in the \href{https://gargle.r-lib.org/articles/request-helper-functions.html#discovery-documents}{Discovery Documents section}
of the gargle vignette \href{https://gargle.r-lib.org/articles/request-helper-functions.html}{Request helper functions}.
}
\examples{
str(head(sm_endpoints(), 3), max.level = 2)
sm_endpoint("secretmanager.projects.secrets.versions.destroy")
sm_endpoint(4)
}
