% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{sm_secret_create}
\alias{sm_secret_create}
\title{Create a Secret}
\usage{
sm_secret_create(
  secret_id,
  project_id = sm_project_get(),
  replication = list(automatic = list()),
  labels = NULL,
  ...
)
}
\arguments{
\item{secret_id}{Required. A unique identifier for the secret within the project.
Must be a string with a maximum length of 255 characters and can contain
uppercase and lowercase letters, numerals, and the hyphen (\code{-}) and
underscore (\verb{_}) characters.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{replication}{Required. The replication policy for the secret data.
Must be a list with either \code{automatic} or \code{user_managed} configuration.}

\item{labels}{Optional. Labels to attach to the secret.}

\item{...}{Additional arguments for methods.}
}
\value{
An \code{sm_secret} object representing the created secret.
}
\description{
Creates a new Secret containing no SecretVersions.
}
