% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{sm_secret_delete}
\alias{sm_secret_delete}
\alias{sm_secret_delete.character}
\alias{sm_secret_delete.sm_secret}
\title{Delete a Secret}
\usage{
sm_secret_delete(secret, project_id = sm_project_get(), ...)

\method{sm_secret_delete}{character}(secret, project_id = sm_project_get(), ...)

\method{sm_secret_delete}{sm_secret}(secret, project_id = sm_project_get(), ...)
}
\arguments{
\item{secret}{The secret to delete. Can be a secret ID (character string)
or an existing \code{sm_secret} object.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{...}{Additional arguments for methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Deletes a Secret and all of its versions.
}
