% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{sm_secret_get}
\alias{sm_secret_get}
\alias{sm_secret_get.character}
\alias{sm_secret_get.sm_secret}
\title{Get Secret Metadata}
\usage{
sm_secret_get(x, project_id = sm_project_get(), ...)

\method{sm_secret_get}{character}(x, project_id = sm_project_get(), ...)

\method{sm_secret_get}{sm_secret}(x, project_id = sm_project_get(), ...)
}
\arguments{
\item{x}{The identifier for the secret. Can be a secret ID (character string)
or an existing \code{sm_secret} object to refresh its metadata.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{...}{Additional arguments for methods.}
}
\description{
Retrieves metadata for a specific Secret Manager secret.
}
