% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret.R
\name{sm_secret_ls.character}
\alias{sm_secret_ls.character}
\title{List Secrets in a Project}
\usage{
\method{sm_secret_ls}{character}(project_id = sm_project_get(), filter = NULL, ...)
}
\arguments{
\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{filter}{Optional. A filter string, adhering to Secret Manager's
\href{https://cloud.google.com/secret-manager/docs/filtering}{List-operation filtering rules}.}

\item{...}{Additional arguments for methods.}
}
\description{
Lists metadata for all Secrets in a given Google Cloud Project.
}
