% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_version.R
\name{sm_secret_version_add}
\alias{sm_secret_version_add}
\alias{sm_secret_version_add.character}
\alias{sm_secret_version_add.sm_secret}
\title{Add a Secret Version}
\usage{
sm_secret_version_add(secret, payload, project_id = sm_project_get(), ...)

\method{sm_secret_version_add}{character}(secret, payload, project_id = sm_project_get(), ...)

\method{sm_secret_version_add}{sm_secret}(secret, payload, project_id = sm_project_get(), ...)
}
\arguments{
\item{secret}{The secret to add a version to. Can be a secret ID (character string)
or an existing \code{sm_secret} object.}

\item{payload}{The secret data to store. Will be base64 encoded.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{...}{Additional arguments for methods.}
}
\value{
An \code{sm_secret_version} object representing the new version.
}
\description{
Adds a new version to an existing Secret.
}
