% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_version.R
\name{sm_secret_version_delete}
\alias{sm_secret_version_delete}
\alias{sm_secret_version_delete.character}
\alias{sm_secret_version_delete.sm_secret}
\title{Delete a Secret Version}
\usage{
sm_secret_version_delete(
  secret,
  version_id,
  project_id = sm_project_get(),
  ...
)

\method{sm_secret_version_delete}{character}(
  secret,
  version_id,
  project_id = sm_project_get(),
  ...
)

\method{sm_secret_version_delete}{sm_secret}(
  secret,
  version_id,
  project_id = sm_project_get(),
  ...
)
}
\arguments{
\item{secret}{The secret containing the version. Can be a secret ID (character string)
or an existing \code{sm_secret} object.}

\item{version_id}{The version ID to delete.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{...}{Additional arguments for methods.}
}
\value{
Invisibly returns \code{NULL}.
}
\description{
Deletes a Secret Version.
}
