% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secret_version.R
\name{sm_secret_version_get}
\alias{sm_secret_version_get}
\alias{sm_secret_version_get.character}
\alias{sm_secret_version_get.sm_secret}
\title{Get Secret Version Metadata}
\usage{
sm_secret_version_get(secret, version_id, project_id = sm_project_get(), ...)

\method{sm_secret_version_get}{character}(secret, version_id, project_id = sm_project_get(), ...)

\method{sm_secret_version_get}{sm_secret}(secret, version_id, project_id = sm_project_get(), ...)
}
\arguments{
\item{secret}{The secret containing the version. Can be a secret ID (character string)
or an existing \code{sm_secret} object.}

\item{version_id}{The version ID to get. Can be "latest" to get the latest version.}

\item{project_id}{The Google Cloud Project ID. Defaults to \code{sm_project_get()}.}

\item{...}{Additional arguments for methods.}
}
\value{
An \code{sm_secret_version} object representing the version metadata.
}
\description{
Gets metadata for a specific Secret Version.
}
