% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request.R
\name{secretmanager_request_make}
\alias{secretmanager_request_make}
\title{Make a request for the Google Secret Manager API}
\usage{
secretmanager_request_make(x, ...)
}
\arguments{
\item{x}{List, holding the components for an HTTP request, presumably created
with \code{\link[=secretmanager_request_generate]{secretmanager_request_generate()}}. Should contain the \code{method}, \code{url}, \code{body},
and \code{token}.}

\item{...}{Optional arguments passed through to the HTTP method.}
}
\value{
An \code{httr::response} object.
}
\description{
This function serves as a thin wrapper around \code{gargle::request_make()}.
It adds the package's user agent and executes the HTTP request, including
handling retries. The input \code{x} is typically created with
\code{\link[=secretmanager_request_generate]{secretmanager_request_generate()}} and the output should then be processed
with \code{\link[gargle:response_process]{gargle::response_process()}}.
}
